slider is synced, need to move seek function to store so other client's players will update as well

+++ backend/app/database/room.py
@@ -8,7 +8,7 @@ class Room(db.Model):
     name = db.Column(db.String(64), index=True, unique=False)
     video_id = db.Column(db.String(64), index=False, unique=False)
     video_url = db.Column(db.String(64), index=False, unique=False)
-    video_time = db.Column(db.Integer, index=False, unique=False)
+    video_time = db.Column(db.Float, index=False, unique=False)
     # Can be PLAYING, PAUSED
     video_state = db.Column(db.String(16), index=False, unique=False)
     valid_video_states = ["PLAYING", "PAUSED"]
